+++ lib/src/extensions/map_extensions.dart
@@ -1,5 +1,7 @@
 extension on Map<String, dynamic> {
   String joinWithMap(String separator, {String keyValueSeparator = '='}) {
-    return entries.map((entry) => "${entry.key}$keyValueSeparator${entry.value}").join(separator);
+    return entries
+        .map((entry) => "${entry.key}$keyValueSeparator${entry.value}")
+        .join(separator);
   }
 }