+++ lib/src/widgets/double_press_back_widget.dart
@@ -1,18 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:nb_utils/nb_utils.dart';
 
-DateTime? _currentBackPressTime;
-
 /// A widget that handles double press back navigation.
-class DoublePressBackWidget extends StatelessWidget {
+///
+/// This widget allows users to navigate back by pressing the back button
+/// twice within a 1-second window. It displays a customizable message
+/// prompting the user for the second press.
+class DoublePressBackWidget extends StatefulWidget {
   /// The child widget to display.
   final Widget child;
 
   /// The message to display when prompting the user to double press back.
   final String? message;
 
-  /// Callback function that gets called on willPop.
-  final WillPopCallback? onWillPop;
+  /// Callback function that gets called on pop confirmation (double press).
+  final VoidCallback? onWillPop; // Changed type to VoidCallback
 
   DoublePressBackWidget({
     super.key,
@@ -22,22 +24,36 @@ class DoublePressBackWidget extends StatelessWidget {
   });
 
   @override
+  State<DoublePressBackWidget> createState() => _DoublePressBackWidgetState();
+}
+
+class _DoublePressBackWidgetState extends State<DoublePressBackWidget> {
+  DateTime? _currentBackPressTime;
+
+  @override
   Widget build(BuildContext context) {
-    return WillPopScope(
-      child: child,
-      onWillPop: () {
-        DateTime now = DateTime.now();
-
-        onWillPop?.call();
-        if (_currentBackPressTime == null ||
-            now.difference(_currentBackPressTime!) > Duration(seconds: 2)) {
-          _currentBackPressTime = now;
-          toast(message ?? 'Press back again to exit');
-
-          return Future.value(false);
+    return PopScope(
+      // Allow pop only after 2 seconds since last press
+      canPop: _currentBackPressTime != null,
+      onPopInvoked: (bool didPop) async {
+        if (didPop) {
+          // User confirmed with double press within 2 seconds
+          widget.onWillPop?.call(); // Call the user-defined callback
+          Navigator.pop(context); // Assuming this triggers back navigation
+          return;
         }
-        return Future.value(true);
+
+        _currentBackPressTime = DateTime.now();
+        setState(() {});
+        toast(widget.message ?? 'Press back again to exit');
+
+        // Start a timer to reset state after 2 seconds
+        Future.delayed(const Duration(seconds: 2), () {
+          _currentBackPressTime = null;
+          setState(() {});
+        });
       },
+      child: widget.child,
     );
   }
 }